pyproject.toml
[tool.poetry]
name = "barista"
version = "0.1.0"
description = "A Python and Jsonnet framework for handling espanso configurations"
authors = ["Jeppe Klitgaard <hi@jeppe.science>"]
license = "MIT"

[tool.poetry.scripts]
barista = "barista.__main__:cli"

[tool.poetry.dependencies]
python = "^3.10"
jsonnet = "^0.17.0"
PyYAML = "^6.0"
click = "^8.0.3"
# TODO: Undo git once pydantic v2 has been released
pydantic = {git = "https://github.com/samuelcolvin/pydantic"}

[tool.poetry.dev-dependencies]
mypy = "^0.920"
black = "^21.9b0"
isort = "^5.9.3"
flake8 = "^4.0.1"
types-PyYAML = "^6.0.1"
poethepoet = "^0.11.0"
ipykernel = "^6.6.0"
pytest = "^6.2.5"

# Mypy
[tool.mypy]
python_version = "3.10"
strict = true
show_error_codes = true
plugins = [
    "pydantic.mypy",
]

[[tool.mypy.overrides]]
module = [
    "_jsonnet.*",
]
ignore_missing_imports = true

# Black
[tool.black]
line-length = 100
target-version = ["py310"]

# Isort
[tool.isort]
profile = "black"
skip_gitignore = true
line_length = 100

# Tasks
[tool.poe.tasks]
flake8 = "flake8 ./barista"
mypy = "mypy ./barista"
black_check = "black --check ."
isort_check = "isort --check-only ."

black = "black ."
isort = "isort ."

jsonnetfmt = "jsonnetfmt -i **/*.jsonnet"

lint = [
    "flake8",
    "black_check",
    "isort_check",
    "mypy",
]


fix = [
    "black",
    "isort",
    "jsonnetfmt"
]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
